pcb: remove redundant fragment of usb_dp
On a wire connecting the usb_dp to the header pin there was a redundant
fragment that was going on top of the actual wire connecting the whole
net. A small part of this wire had an offset making the wire ticker in
one place as well.

I noticed that by pure luck by manually going over the wire (after
pressing I). Unfortunately I don't know any method of automatically
finding such problems.

Note that in order to keep this change as small as possible I did not
run the refill of the bottom layer even though it could flow a little
more now that the thicker part of a wire is gone.