footprints: fix courtyard of wson8-8x6mm
One edge (left one) of the courtyard box in this package was drawn on
the wrong layer (Dwgs.User instead of F.CrtYd) like others. While
hard/impossible to see by an eye, for the Kicad this was unclosed shape
making it complain in DRC if courtyard checks are enabled.

Note that in order to make this change small and easy to merge in case
of conflicts, I did not include the change on the kicad_pcb file after
updating the footprint. You have to click on the "Update footprint from
library" in the properties window of U2 yourself.